Added the new classes and changed the Resumes to extend the new class
